+++ scripts/update_epoch_ada_pots.sh
@@ -10,7 +10,7 @@ set -e
 source config.pg
 source functions.sh
 
-TNAME="epoch_param"
+TNAME="ada_pots"
 
 BQ_EPOCH_NO=$1
 PG_EPOCH_NO=$2
@@ -27,31 +27,18 @@ else
   echo "   loading for epoch $EPOCH_NO"
 fi
 
-CSVNAME="update_epoch_param-q1"
+CSVNAME="update_${TNAME}-q1"
 if [ -e "${CSVNAME}.csv" ]; then rm -f "${CSVNAME}.csv"; fi
-SCHEMA="epoch_param"
+SCHEMA="${TNAME}"
 DATASETID="iog-data-analytics:db_sync"
 
-## 1 insert epoch into table: tmp_epoch_param_1
-TMPTBL="tmp_epoch_param_1"
+## 1 insert epoch into temporary table
+TMPTBL="tmp_${TNAME}_1"
 Q="
-    SELECT epoch_no, min_fee_a, min_fee_b,
-         max_block_size, max_tx_size, max_bh_size,
-         key_deposit, pool_deposit, max_epoch,
-         optimal_pool_count, influence, 
-         monetary_expand_rate, treasury_growth_rate,
-         decentralisation,
-         extra_entropy,
-         protocol_major, protocol_minor,
-         min_utxo_value, min_pool_cost,
-         nonce,
-         coins_per_utxo_size, cost_model,
-         price_mem, price_step,
-         max_tx_ex_mem, max_tx_ex_steps,
-         max_block_ex_mem, max_block_ex_steps,
-         max_val_size,
-         collateral_percent, max_collateral_inputs
-  FROM analytics.vw_bq_epoch_param
+  SELECT epoch_no, slot_no,
+         treasury, reserves, rewards, utxo,
+         deposits, fees
+  FROM public.${TNAME}
   WHERE epoch_no = ${EPOCH_NO}"
 NREAD=$(pg_query_to_csv "${Q}" "$CSVNAME")